feat: leave method and optimizations

+++ app/controllers/parking_spots_controller.rb
@@ -22,21 +22,37 @@ class ParkingSpotsController < ApplicationController
   def edit
   end
 
-  def update
-  end
-
   def destroy
     parking_spot = ParkingSpot.find(params[:id])
-    parking_spot.destroy
-    redirect_to parking_spots_path, notice: "Parking spot removed."
+    if parking_spot.car_id != nil
+      car = Car.find(parking_spot.car_id)
+      parking_spot.destroy
+      car.destroy
+    end
+    redirect_to root_path
   end
 
   def park
     empty_spot = ParkingSpot.find_empty
-    car = Car.create(license: params[:license], color: params[:color])
-    if car.save
-      empty_spot.update(car: car)
-      redirect_to parking_spots_path
+    if !empty_spot.respond_to?("length")
+      car = Car.create(license: params[:license], color: params[:color])
+      if car.save
+        empty_spot.update(car: car)
+        redirect_to root_path
+      end
+    else
+      redirect_to root_path, notice: "Parking lot full"
     end
+
+  end
+
+  def leave
+    parking_spot = ParkingSpot.find(params[:id])
+    if parking_spot.car_id != nil
+      car = Car.find(parking_spot.car_id)
+      parking_spot.update(car_id: nil)
+      car.destroy
+    end
+    redirect_to root_path
   end
 end
